created generic class for attacks
