+++ Assets/Scripts/Attacks/Attack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dd213eff9372ed747840ea7760a3f7f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
