+++ Assets/Scripts/Attacks/AttackHoming.cs
@@ -1,31 +1,8 @@
 using System.Collections;
 using UnityEngine;
 
-public class AttackHoming : MonoBehaviour
+public class AttackHoming : Attack
 {
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        if (transform.position.y < 0f || transform.position.y > Screen.height || transform.position.x > Screen.width || transform.position.x < 0)
-        {
-            Destroy(gameObject);
-        }
-    }
-
-    private void OnTriggerEnter2D(Collider2D collision)
-    {
-        if (collision.tag == Constants.PLAYER_TAG)
-        {
-            Destroy(gameObject);
-        }
-    }
-
     public static void ShootOnDemand(GameObject enemy, GameObject missile, float shootSpeed, Transform shootPos, Transform playerPos) 
     {
         GameObject newMissile = Instantiate(missile, shootPos.position, Quaternion.identity) as GameObject;
